requirements.txt
# Core dependencies
dspy-ai>=2.4.0
boto3>=1.34.0
python-docx>=1.1.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Date handling
python-dateutil>=2.8.0

# Optional: For enhanced functionality
requests>=2.31.0
typing-extensions>=4.5.0